# Improve serde performance for Long blocks