--- a/presto-main/src/test/java/io/prestosql/execution/buffer/TestingPagesSerdeFactory.java
+++ b/presto-main/src/test/java/io/prestosql/execution/buffer/TestingPagesSerdeFactory.java
@@ -31,7 +31,12 @@ public class TestingPagesSerdeFactory
     public TestingPagesSerdeFactory()
     {
         // compression should be enabled in as many tests as possible
-        super(createTestMetadataManager().getBlockEncodingSerde(), true);
+        this(createTestMetadataManager().getBlockEncodingSerde(), true);
+    }
+
+    public TestingPagesSerdeFactory(BlockEncodingSerde blockEncodingSerde, boolean compressionEnabled)
+    {
+        super(blockEncodingSerde, compressionEnabled);
     }
 
     public static PagesSerde testingPagesSerde()