--- a/presto-orc/src/test/java/io/prestosql/orc/OrcTester.java
+++ b/presto-orc/src/test/java/io/prestosql/orc/OrcTester.java
@@ -105,7 +105,9 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.Properties;
 import java.util.Set;
+import java.util.stream.IntStream;
 
+import static com.google.common.collect.ImmutableList.toImmutableList;
 import static com.google.common.collect.Iterators.advance;
 import static com.google.common.collect.Lists.newArrayList;
 import static io.airlift.slice.Slices.utf8Slice;
@@ -600,6 +602,34 @@ public class OrcTester
                 newSimpleAggregatedMemoryContext(),
                 initialBatchSize,
                 RuntimeException::new);
+    }
+
+    public static void writeOrcPages(File outputFile, CompressionKind compression, List<Type> types, Iterator<Page> pages, OrcWriterStats stats)
+            throws Exception
+    {
+        List<String> columnNames = IntStream.range(0, types.size())
+                .mapToObj(i -> "test" + i)
+                .collect(toImmutableList());
+
+        OrcWriter writer = new OrcWriter(
+                new OutputStreamOrcDataSink(new FileOutputStream(outputFile)),
+                columnNames,
+                types,
+                OrcType.createRootOrcType(columnNames, types),
+                compression,
+                new OrcWriterOptions(),
+                false,
+                ImmutableMap.of(),
+                true,
+                BOTH,
+                stats);
+
+        while (pages.hasNext()) {
+            writer.write(pages.next());
+        }
+
+        writer.close();
+        writer.validate(new FileOrcDataSource(outputFile, READER_OPTIONS));
     }
 
     public static void writeOrcColumnPresto(File outputFile, CompressionKind compression, Type type, Iterator<?> values, OrcWriterStats stats)