--- a/presto-spi/src/main/java/io/prestosql/spi/block/LongArrayBlock.java
+++ b/presto-spi/src/main/java/io/prestosql/spi/block/LongArrayBlock.java
@@ -13,6 +13,8 @@
  */
 package io.prestosql.spi.block;
 
+import io.airlift.slice.Slice;
+import io.airlift.slice.Slices;
 import org.openjdk.jol.info.ClassLayout;
 
 import javax.annotation.Nullable;
@@ -262,6 +264,11 @@ public class LongArrayBlock
         sb.append("positionCount=").append(getPositionCount());
         sb.append('}');
         return sb.toString();
+    }
+
+    Slice getValuesSlice()
+    {
+        return Slices.wrappedLongArray(values, arrayOffset, positionCount);
     }
 
     private void checkReadablePosition(int position)