--- a/presto-spi/src/test/java/io/prestosql/spi/block/BlockTestUtils.java
+++ b/presto-spi/src/test/java/io/prestosql/spi/block/BlockTestUtils.java
@@ -1,0 +1,31 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.prestosql.spi.block;
+
+import io.prestosql.spi.type.Type;
+
+import static io.prestosql.spi.block.TestingSession.SESSION;
+import static org.testng.Assert.assertEquals;
+
+public class BlockTestUtils
+{
+    private BlockTestUtils() {}
+
+    public static void assertBlockEquals(Type type, Block actual, Block expected)
+    {
+        for (int position = 0; position < actual.getPositionCount(); position++) {
+            assertEquals(type.getObjectValue(SESSION, actual, position), type.getObjectValue(SESSION, expected, position));
+        }
+    }
+}