… presto-orc/pom.xml
… presto-spi/src/test/java/io/prestosql/spi/block/TestDictionaryBlockEncoding.java
… presto-spi/src/test/java/io/prestosql/spi/block/TestLongArrayBlockEncoding.java
… presto-spi/src/test/java/io/prestosql/spi/block/TestVariableWidthBlockEncoding.java