```
AFTER (no null)

Benchmark                                    Mode  Cnt  Score   Error  Units
BenchmarkBlockSerde.deserializeLongNoNull    avgt   60  1,125 ± 0,011  ns/op

BEFORE (no null)

Benchmark                                    Mode  Cnt  Score   Error  Units
BenchmarkBlockSerde.deserializeLongNoNull    avgt   60  2,548 ± 0,023  ns/op

AFTER (no null)

Benchmark                                    Mode  Cnt  Score   Error  Units
BenchmarkBlockSerde.deserializeLineitem      avgt   60  1,212 ± 0,015  ns/op
BenchmarkBlockSerde.deserializeLongWithNull  avgt   60  2,500 ± 0,038  ns/op
BenchmarkBlockSerde.serializeLineitem        avgt   60  3,645 ± 0,039  ns/op
BenchmarkBlockSerde.serializeLongWithNull    avgt   60  3,351 ± 0,016  ns/op

BEFORE (no null)

Benchmark                                    Mode  Cnt  Score   Error  Units
BenchmarkBlockSerde.deserializeLineitem      avgt   60  1,219 ± 0,016  ns/op
BenchmarkBlockSerde.deserializeLongWithNull  avgt   60  5,524 ± 0,062  ns/op
BenchmarkBlockSerde.serializeLineitem        avgt   60  3,671 ± 0,041  ns/op
BenchmarkBlockSerde.serializeLongWithNull    avgt   60  8,844 ± 0,086  ns/op
```