@sopel39 2020-11-12
addressed comments, improved serialization and de-serialization for long blocks with nulls too